+++ R/2_onLoad.R
@@ -1,9 +1,4 @@
 .onLoad <- function(libname, pkgname) {
-  if (base::requireNamespace("rstudioapi", quietly = TRUE)) {
-    CONFIG$rstudio <- rstudioapi::isAvailable()
-  } else if (interactive() & Sys.getenv("RSTUDIO") == 1) {
-    warning("Install 'rstudioapi' package to enable breakpoint debugging")
-  }
 
   invisible()
 }