+++ tests/testthat/test_initialize_project.R
@@ -2,11 +2,11 @@ context("initialize")
 
 test_that("Create org::project$results_today", {
   initialize_project(
+    env = .GlobalEnv,
     home = tempdir(),
     results = tempdir(),
     raw = tempdir(),
-    folders_to_be_sourced = NULL,
-    create_folders = TRUE
+    folders_to_be_sourced = NULL
   )
 
   testthat::expect_equal(TRUE, dir.exists(org::project$results_today))
@@ -15,6 +15,7 @@ test_that("Create org::project$results_today", {
 test_that("Error due to multiple non-existed folders", {
   testthat::expect_error(
     initialize_project(
+      env = .GlobalEnv,
       home = c("dfsdfoij323423", "sdfd232323"),
       results = tempdir(),
       raw = tempdir()
@@ -25,11 +26,11 @@ test_that("Error due to multiple non-existed folders", {
 
 test_that("Works due to multiple non-existed folders", {
   initialize_project(
+    env = .GlobalEnv,
     home = c(tempdir(), "sdfd232323"),
     results = tempdir(),
     raw = tempdir(),
-    folders_to_be_sourced = NULL,
-    create_folders = TRUE
+    folders_to_be_sourced = NULL
   )
 
   testthat::expect_equal(TRUE, dir.exists(org::project$results_today))
@@ -38,11 +39,11 @@ test_that("Works due to multiple non-existed folders", {
 
 test_that("computer_id identifying correct order", {
   initialize_project(
+    env = .GlobalEnv,
     home = c("sdfd232323", tempdir()),
     results = tempdir(),
     raw = tempdir(),
-    folders_to_be_sourced = NULL,
-    create_folders = TRUE
+    folders_to_be_sourced = NULL
   )
 
   testthat::expect_equal(2, org::project$computer_id)
@@ -54,10 +55,10 @@ test_that("Sources multiple code folders that do exist", {
 
   testthat::expect_message(
     initialize_project(
+      env = .GlobalEnv,
       home = tempdir(),
       raw = tempdir(),
-      folders_to_be_sourced = c("x1", "y1"),
-      create_folders = TRUE
+      folders_to_be_sourced = c("x1", "y1")
     ),
     "*Sourcing all code inside*"
   )
@@ -68,10 +69,10 @@ test_that("Sources multiple code folders that dont exist", {
   dir.create(tempdir())
   testthat::expect_warning(
     initialize_project(
+      env = .GlobalEnv,
       home = tempdir(),
       raw = tempdir(),
-      folders_to_be_sourced = c("x2", "y2"),
-      create_folders = TRUE
+      folders_to_be_sourced = c("x2", "y2")
     ),
     "*Creating it now."
   )